texk/ps2pkm: Minor bug fix. Avoid use of LONG (int for Unix but defined as long by MinGW)
git-svn-id: svn://tug.org/texlive/trunk/Build/source@35365 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/ps2pkm/curves.c
@@ -182,7 +182,7 @@ the starting values.  If this overflows, a 'long', we are in trouble:
 #if defined(BITS)
 #undef BITS
 #endif
-#define  BITS         (sizeof(LONG)*8)
+#define  BITS         (sizeof(int32_t)*8)
 #define  HIGHTEST(p)  (((p)>>(BITS-4)) != 0)  /* includes sign bit */
 #define  TOOBIG(xy)   ((xy < 0) ? HIGHTEST(-xy) : HIGHTEST(xy))
